wip: begin abandon CDKTF: AWS EKS
